app/pyproject.toml
[project]
name = "cineplexx-rss"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
  "playwright",
  "python-dotenv",
  "redis>=5",
]

[project.scripts]
cineplexx-rss = "cineplexx_rss.main:main"
